Periodically resynchronize time.
We are currently using the device timestamps to determine the
"local" time that should be attached to the reading.  This
works for a while, but since the crystal on the IMU and the
crystal on the local machine are separate, they eventually
drift apart.  Fix this by periodically resynchronizing between
the local time and the device time.  There will always be
some difference between them, but with the algorithm implemented
here we know that that difference will be <= 1000/imu_rate (and
usually will be a lot less).

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>